sql/tree: move role validation test to correct package
Release note: None
