+++ pkg/sql/BUILD.bazel
@@ -451,7 +451,6 @@ go_test(
         "copy_in_test.go",
         "copy_test.go",
         "crdb_internal_test.go",
-        "create_role_test.go",
         "create_stats_test.go",
         "create_test.go",
         "database_test.go",
@@ -586,7 +585,6 @@ go_test(
         "//pkg/sql/opt/exec/explain",
         "//pkg/sql/parser",
         "//pkg/sql/pgwire/pgcode",
-        "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/pgwire/pgwirebase",
         "//pkg/sql/physicalplan",
         "//pkg/sql/querycache",
