+++ pkg/sql/sem/tree/BUILD.bazel
@@ -203,6 +203,7 @@ go_test(
         "parse_array_test.go",
         "placeholders_test.go",
         "pretty_test.go",
+        "role_spec_test.go",
         "table_name_test.go",
         "time_test.go",
         "timeconv_test.go",
@@ -227,6 +228,8 @@ go_test(
         "//pkg/sql/opt/optbuilder",
         "//pkg/sql/opt/xform",
         "//pkg/sql/parser",
+        "//pkg/sql/pgwire/pgcode",
+        "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/randgen",
         "//pkg/sql/sem/builtins",
         "//pkg/sql/sessiondata",
